Rename settings package to asb_config
